use lump instead of page as page number

+++ z_zone.h
@@ -495,7 +495,7 @@ void __far Z_SetOverlay(int8_t wipeId);
 // EMS page frame stuff
 void __far Z_QuickMapPageFrame(uint8_t pageframeindex, uint8_t pagenumber);
 void __far Z_QuickMapSFXPageFrame(uint8_t pagenumber);
-void __far Z_QuickMapWADPageFrame(uint8_t pagenumber);
+void __far Z_QuickMapWADPageFrame(int16_t pagenumber);
 
 void __far Z_SavePageFrameState();
 void __far Z_RestorePageFrameState();
